fix: Cambiar tipo de dato de empleadoId a Short en NominaController y NominaRepository para mejorar la consistencia de tipos feat: Añadir funciones getEmployee y getDocumentById en EmployeeHelper para facilitar la obtención de datos de empleados refactor: Actualizar texto de opción en SeveranceWithdrawScreen para corregir la acentuación en "Terminación de contrato"

+++ BACKEND_PAYROLL/src/main/java/co/edu/unbosque/backend_payroll/controller/NominaController.java
@@ -24,7 +24,7 @@ public class NominaController {
     }
 
     @GetMapping("/empleado/{empleadoId}")
-    public ResponseEntity<List<NominaProjection>> getNominasByEmpleado(@PathVariable Integer empleadoId) {
+    public ResponseEntity<List<NominaProjection>> getNominasByEmpleado(@PathVariable Short empleadoId) {
         List<NominaProjection> nominas = nominaService.getNominasByEmpleadoId(empleadoId);
         return ResponseEntity.ok(nominas);
     }
